+++ CHANGELOG.md
@@ -9,13 +9,16 @@
 ## ✨ What's New
 - AVP now requires AzureLib `3.0.32`.
 - Marines can now be given armor items.
+- Marines can now use armor items:
   - Marines will now automatically figure out what armor is best to wear depending on their situation.
   - By default, marines will equip whatever armor offers the best protection. This accounts for defense points, toughness and protection enchantment(s).
   - If underwater, marines will equip a pressure suit or MK50 to breathe infinitely.
     - Marines will only equip these armors if they have a full set of either armor type.
 - Marines can now be given potion items.
+- Marines can now use potion items:
   - If on fire, marines will now consider drinking a fire resistance potion.
   - If underwater, marines will now consider drinking a water breathing potion if they are close to drowning.
+  - If not at full health, marines will now consider drinking an instant health potion.
 
 ## ♻️ Changes
 - N/A